+++ src/test/java/com/n3twork/dynamap/DynamapTxTest.java
@@ -22,11 +22,9 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.dynamodbv2.local.embedded.DynamoDBEmbedded;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.n3twork.dynamap.*;
 import com.n3twork.dynamap.test.PlayerBean;
 import com.n3twork.dynamap.test.PlayerUpdates;
-import com.n3twork.dynamap.tx.ReadTx;
-import com.n3twork.dynamap.tx.WriteConditionCheck;
-import com.n3twork.dynamap.tx.WriteTx;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
